changed datatype of variable 'kerning' from int to float.

+++ demo-glyph.c
@@ -146,7 +146,7 @@ void add_text( vertex_buffer_t * buffer, texture_font_t * font,
         texture_glyph_t *glyph = texture_font_get_glyph( font, text[i] );
         if( glyph != NULL )
         {
-            int kerning = 0;
+            float kerning = 0.0f;
             if( i > 0)
             {
                 kerning = texture_glyph_get_kerning( glyph, text[i-1] );
